refactor: consolidate git submodule handling into git_submodule.py
Submodule-related logic was spread across three places: GitLocalRepo
methods in git.py (submodules(), _get_submodule_urls(),
_filter_submodules_by_src(), _move_src_folder_up() and friends), the
standalone git_gitlinks.py (orphan-gitlink stripping added for #1380),
and the Submodule dataclass in git_types.py.

Move all of it into one dfetch/vcs/git_submodule.py:
- Submodule dataclass (moved from git_types.py; CheckoutOptions stays,
  it isn't submodule-specific)
- everything from git_gitlinks.py (gitlink_paths, declared_submodule_paths,
  drop_orphan_gitlinks)
- submodule url/branch resolution (get_submodule_urls, ensure_abs_url)
- src/ignore filtering and promotion of a fetched submodule tree
  (apply_src_and_ignore, filter_submodules_by_src, remove_empty_parents,
  move_src_folder_up and its helpers)

None of the moved code touched GitLocalRepo instance state, so it
translates directly into free functions. GitLocalRepo.submodules()
stays in git.py as a thin orchestrator: it's the one piece that
genuinely needs GitRemote/GitLocalRepo (to resolve a submodule's branch
from its own remote or local history), so keeping it there avoids a
circular import between the two modules while git.py imports
git_submodule normally for everything else.

Test files follow: tests/test_git_gitlinks.py is folded into a new
tests/test_git_submodule.py along with the filter/move tests that used
to live in test_git_vcs.py, mirroring the module split.

No behavioral change: full pytest (705) and non-SVN behave suites pass
identically before and after.
