+++ dfetch/vcs/git_types.py
@@ -5,19 +5,6 @@ from dataclasses import dataclass
 
 
 @dataclass
-class Submodule:
-    """Information about a submodule."""
-
-    name: str
-    toplevel: str
-    path: str
-    sha: str
-    url: str
-    branch: str
-    tag: str
-
-
-@dataclass
 class CheckoutOptions:
     """Options for checking out a specific version from a remote git repository."""
