+++ tests/test_git_vcs.py
@@ -17,7 +17,7 @@ from dfetch.vcs.git import (
     GitRemote,
     _build_git_ssh_command,
 )
-from dfetch.vcs.git_types import CheckoutOptions, Submodule
+from dfetch.vcs.git_types import CheckoutOptions
 
 # ---------------------------------------------------------------------------
 # unique_parent_dirs  (dfetch.util.util)
@@ -78,146 +78,6 @@ def test_build_hash_args(name, old_hash, new_hash, expected):
 )
 def test_build_ignore_args(name, ignore, expected):
     assert GitLocalRepo._build_ignore_args(ignore) == expected
-
-
-# ---------------------------------------------------------------------------
-# GitLocalRepo._move_src_folder_up — path-traversal guards
-# ---------------------------------------------------------------------------
-
-
-def test_move_src_folder_up_rejects_absolute_src(tmp_path):
-    """An absolute src pattern must be rejected without touching the filesystem."""
-    with patch("dfetch.vcs.git.move_directory_contents") as mock_move:
-        with patch("dfetch.vcs.git.os.getcwd", return_value=str(tmp_path)):
-            GitLocalRepo._move_src_folder_up("my-remote", "/etc")
-    mock_move.assert_not_called()
-
-
-def test_move_src_folder_up_rejects_traversal_src(tmp_path):
-    """A src pattern that resolves outside the repo root must be skipped."""
-    outside = tmp_path.parent / "outside"
-    outside.mkdir()
-    (outside / "secret.txt").write_text("data")
-
-    with patch("dfetch.vcs.git.move_directory_contents") as mock_move:
-        with patch("dfetch.util.util.glob.glob", return_value=[str(outside)]):
-            with patch("dfetch.vcs.git.os.getcwd", return_value=str(tmp_path)):
-                GitLocalRepo._move_src_folder_up("my-remote", "../outside")
-    mock_move.assert_not_called()
-
-
-# ---------------------------------------------------------------------------
-# GitLocalRepo._filter_submodules_by_src
-# ---------------------------------------------------------------------------
-
-
-def _make_submodule(path: str, url: str = "some-url") -> Submodule:
-    return Submodule(
-        name=path,
-        toplevel="",
-        path=path,
-        sha="abc123",
-        url=url,
-        branch="master",
-        tag="",
-    )
-
-
-def test_filter_submodules_ancestor_of_src_not_removed(tmp_path, monkeypatch):
-    """A submodule whose path is an ancestor of src must not be deleted.
-
-    When src='apps/myapp' and a submodule exists at 'apps', the old logic
-    incorrectly added 'apps' to to_remove (because Path('apps/myapp').is_relative_to('apps')
-    was True with sub_top='apps').  The fix checks the full submodule.path so that an
-    ancestor submodule is skipped and _move_src_folder_up can promote its content.
-    """
-    (tmp_path / "apps" / "myapp").mkdir(parents=True)
-    (tmp_path / "apps" / "myapp" / "README.md").write_text("content")
-    (tmp_path / "outside").mkdir()
-    (tmp_path / "outside" / "file.txt").write_text("content")
-
-    monkeypatch.chdir(tmp_path)
-    repo = GitLocalRepo(str(tmp_path))
-    result = repo._filter_submodules_by_src(
-        "remote-url",
-        "apps/myapp",
-        [_make_submodule("apps"), _make_submodule("outside")],
-    )
-
-    assert (
-        tmp_path / "README.md"
-    ).exists(), "apps/myapp content should be promoted to root"
-    assert not (tmp_path / "outside").exists(), "outside/ submodule should be removed"
-    assert not any(
-        s.path == "apps" for s in result
-    ), "ancestor submodule must not appear in result"
-    assert not any(
-        s.path == "outside" for s in result
-    ), "out-of-scope submodule must not appear in result"
-
-
-def test_filter_submodules_disjoint_submodule_removed(tmp_path, monkeypatch):
-    """A submodule whose path is disjoint from src must be removed.
-
-    The fix stores the full submodule.path (not its top-level component) in
-    to_remove, so only the exact submodule directory is deleted.
-    """
-    (tmp_path / "src_folder" / "ext" / "inside").mkdir(parents=True)
-    (tmp_path / "src_folder" / "ext" / "inside" / "README.md").write_text("content")
-    (tmp_path / "other_ext" / "outside").mkdir(parents=True)
-    (tmp_path / "other_ext" / "outside" / "README.md").write_text("content")
-
-    monkeypatch.chdir(tmp_path)
-    repo = GitLocalRepo(str(tmp_path))
-    result = repo._filter_submodules_by_src(
-        "remote-url",
-        "src_folder",
-        [
-            _make_submodule("src_folder/ext/inside"),
-            _make_submodule("other_ext/outside"),
-        ],
-    )
-
-    assert not (
-        tmp_path / "other_ext" / "outside"
-    ).exists(), "other_ext/outside submodule dir should be removed"
-    assert any(
-        s.path == "ext/inside" for s in result
-    ), "inside submodule should be promoted"
-
-
-def test_filter_submodules_sibling_of_src_not_removed(tmp_path, monkeypatch):
-    """A sibling submodule sharing the same top-level dir as src must not destroy src content.
-
-    When src='apps/lib' and submodules exist at both 'apps/lib' (exact match, the src)
-    and 'apps/widget' (sibling, outside src), the old logic used parts[0]='apps' and
-    called safe_rm('apps'), which deleted the entire apps/ directory — including the
-    already-cloned apps/lib content needed by _move_src_folder_up.
-    The fix stores the full submodule.path so only apps/widget is targeted, leaving
-    apps/lib intact for promotion.
-    """
-    (tmp_path / "apps" / "lib").mkdir(parents=True)
-    (tmp_path / "apps" / "lib" / "README.md").write_text("content")
-    (tmp_path / "apps" / "widget").mkdir(parents=True)
-    (tmp_path / "apps" / "widget" / "widget.h").write_text("content")
-
-    monkeypatch.chdir(tmp_path)
-    repo = GitLocalRepo(str(tmp_path))
-    result = repo._filter_submodules_by_src(
-        "remote-url",
-        "apps/lib",
-        [_make_submodule("apps/lib"), _make_submodule("apps/widget")],
-    )
-
-    assert (
-        tmp_path / "README.md"
-    ).exists(), "apps/lib content must be promoted to root"
-    assert not (
-        tmp_path / "apps" / "widget"
-    ).exists(), "sibling apps/widget must be removed"
-    assert any(
-        s.path == "apps/lib" for s in result
-    ), "src submodule should appear in result before final os.path.exists filtering"
 
 
 # ---------------------------------------------------------------------------
